prototype revive in dungeon
ControllerConstants.java - cost to revive
Event.proto - redacting revive in dungeon request and response
CollectEquipEnhancementController.java - userEquipId is a long/biging not int
InsertUtils.java - above
UpdateUtils.java - increasing num revives in user_task

+++ src/main/java/com/lvl6/server/controller/CollectEquipEnhancementController.java
@@ -259,7 +259,7 @@ import com.lvl6.utils.utilmethods.InsertUtils;
     int previousEnhancementPercentage = equipUnderEnhancement.getEnhancementPercentage();
     Timestamp timeOfEnhancement = null;
     Timestamp timeOfSpeedUp = null;
-    int userEquipId = userEquipBuilder.getUserEquipId();
+    long userEquipId = userEquipBuilder.getUserEquipId();
     
     //to prevent possible null pointer exception
     Date startDate = equipUnderEnhancement.getStartTimeOfEnhancement();
